+++ src/vaults/BalanceSheetManager.sol
@@ -140,8 +140,8 @@ contract BalanceSheetManager is
         auth
     {
         uint128 assetId = poolManager.checkedAssetToId(asset, tokenId);
-        sender.sendUpdateHoldingValue(poolId, scId, AssetId.wrap(assetId), pricePerUnit, block.timestamp);
-        emit UpdateValue(poolId, scId, asset, tokenId, pricePerUnit, block.timestamp);
+        sender.sendUpdateHoldingValue(poolId, scId, AssetId.wrap(assetId), pricePerUnit, uint64(block.timestamp));
+        emit UpdateValue(poolId, scId, asset, tokenId, pricePerUnit, uint64(block.timestamp));
     }
 
     /// @inheritdoc IBalanceSheetManager
@@ -231,7 +231,7 @@ contract BalanceSheetManager is
             ITranche(token).mint(address(to), shares);
         }
 
-        sender.sendUpdateShares(poolId, scId, to, pricePerShare, shares, block.timestamp, true);
+        sender.sendUpdateShares(poolId, scId, to, pricePerShare, shares, uint64(block.timestamp), true);
         emit Issue(poolId, scId, to, pricePerShare, shares);
     }
 
@@ -239,7 +239,7 @@ contract BalanceSheetManager is
         address token = poolManager.checkedTranche(poolId.raw(), scId.raw());
         ITranche(token).burn(address(from), shares);
 
-        sender.sendUpdateShares(poolId, scId, from, pricePerShare, shares, block.timestamp, false);
+        sender.sendUpdateShares(poolId, scId, from, pricePerShare, shares, uint64(block.timestamp), false);
         emit Revoke(poolId, scId, from, pricePerShare, shares);
     }
 
@@ -274,10 +274,12 @@ contract BalanceSheetManager is
             }
         }
 
-        sender.sendUpdateHoldingAmount(poolId, scId, assetId, receiver, amount, pricePerUnit, block.timestamp, true, m);
+        sender.sendUpdateHoldingAmount(
+            poolId, scId, assetId, receiver, amount, pricePerUnit, uint64(block.timestamp), true, m
+        );
 
         emit Withdraw(
-            poolId, scId, asset, tokenId, receiver, amount, pricePerUnit, block.timestamp, m.debits, m.credits
+            poolId, scId, asset, tokenId, receiver, amount, pricePerUnit, uint64(block.timestamp), m.debits, m.credits
         );
     }
 
@@ -302,9 +304,13 @@ contract BalanceSheetManager is
         }
 
         escrow.deposit(asset, tokenId, poolId.raw(), scId.raw(), amount);
-        sender.sendUpdateHoldingAmount(poolId, scId, assetId, provider, amount, pricePerUnit, block.timestamp, false, m);
+        sender.sendUpdateHoldingAmount(
+            poolId, scId, assetId, provider, amount, pricePerUnit, uint64(block.timestamp), false, m
+        );
 
-        emit Deposit(poolId, scId, asset, tokenId, provider, amount, pricePerUnit, block.timestamp, m.debits, m.credits);
+        emit Deposit(
+            poolId, scId, asset, tokenId, provider, amount, pricePerUnit, uint64(block.timestamp), m.debits, m.credits
+        );
     }
 
     function _ensureEntries(uint128 amount, Meta calldata m) internal pure {
